DataMapper ORM backed by PDO
This is the first pass at the DataMapper ORM.  This gives us
the ability to store the fields of the data object as and array
in the class instead of relying on migrations and database
structure.

ORMClass::migrate() will automatically create and update tables
as appropriate.  See the test suite for more ideas on how it
works until we get documentation together on it.

We've also made it so that the class loading is more namespace
friendly and allows us to have the same filename in two different
directories as long as they're in separate namespaces.  This
should be one more step into making same class and filenames
throughout the system.