Ability to install via PEAR
Now we have the ability to install via PEAR.  The details aren't totally
figured out yet, but there will be a PEAR channel on
http://pear.silkframework.com.  A /usr/bin/silk command line app will be
installed and be usable over the course of the whole machine.  Plus, the 
libraries only have to be installed once.  There are still a few changes that
need to made in order for the app generation to work as expected, but
this lays the tracks for that to happen.

Also fixed a few bugs with test running, and some other little things.