Added subtable support for table definitions
It allows us to create tables that ORM'd objects need for their
associations, etc. Example would be an object needing a key/value table
for something it's storing.

Use 'type' of 'table' and add a 'fields' key as well, with the fields
just like a normal table definition. It does support recursion if
necessary.