Switch Simpletest to PHPUnit
It's more maintained and used in more places. Plus, it has built in
support for code coverage, and other niceties. The downside is that the
functional testing using Selenium and requires a bit more setup, but I
think it's worth the tradeoff.

PHPUnit will have to be installed separately. It's possible that we
could include it in the Silk PEAR repository somehow, but let's leave
that for something down the road.