Test Fixtures
Ability to setup data beforehand for consistent testing. Only works on
models for the moment, but will add the ability to test against one-off
tables later on. See test/fixtures/fixture.test_suite.php for an idea of
how to set them up.

Also made it so that all database prefixes get test_ prepended to their
names so that we don't have overlap. Fixtures will delete their tables,
but it's up to the people doing it manually to clean up after their done
to not leave a messy database.