Removed old console commands/code
We're now using phake, which is a very simple
library based on ruby's rake. It's not as complex,
which is perfect for creating simple commands. It
also allows us to include commands right in our
extensions and have them magically just show up.

I've converted all the old command line stuff over
to the new syntax and put them into .phake files
in lib/silk/tasks.